Initial commit
A proprietary copright is included. This will change to the GPL once I'm
comfortable distributing the underlying source. Content for pages will
be under a Creative Commons license, unless otherwise specified.

A very primitive version of bin/make-site is included. The README
documents the behavior of this script, though the script does not yet
perform all the tasks that are described.