Added bin/make-chroot
Since the plan is to run Apache under a jail, it seems prudent to have a
script to automatically create the jail. The principle is that jails are
disposable; One should have no qualms destroying a jail and creating a
new one from scratch. To allow this, we must ensure no changes are made
within a jail that cannot be reproduced by this script, or through the
makefile.