Pulled some inlined site-specific variables out to later targets
Specifically, I am working to reduce the amount of code that is
hardcoded to be site-dependent. The hope is that as work continues,
boilerplate living within site directories will be minimized and shifted
to one authoritative location. One prerequisite of this move is
eliminating areas where we hardcode site-dependent variables early in
the workflow.

All that being said, while I can envision broader strokes to eliminate
dependency, I'm not confident that these changes would be beneficial. I
do not want to lose the flexibility I have now by introducing
constraints on how things are configured. So, changes like these will be
incremental, rather than revolutionary.