Initial attempt at enabling SSL
I'm wondering if it would have been better to just force SSL
configurations to live in a different website. The reason I went with
the one-website approach is because it seems like it will be common to
have secure and non-secure sites depend on the same assets, the same
content, and so forth. Where encryption is used is more of an
implementation detail rather than a strong division. While having one
website use two ports complicates things a bit, I believe have them
united in one website will pay dividends in the long run.