Lots of movement regarding SSL certs
I had wanted to get an authoritative CA certificate created that I could
reuse everywhere. However, it seems that there's no easy way to
accomplish that, at least in Firefox. The easiest way is just to bite
the bullet and add exceptions for each site that I create. It's also
safer, so there's that.